requirements.txt
aiogram
APScheduler
python-dotenv